Feature: support Android 12

+++ app/src/main/java/com/xposed/hook/config/Constants.java
@@ -6,6 +6,6 @@ public class Constants {
 
     public static final String DEFAULT_LATITUDE = "34.752600";
     public static final String DEFAULT_LONGITUDE = "113.662000";
-    public static final int DEFAULT_LAC = -1;
-    public static final int DEFAULT_CID = -1;
+    public static final long DEFAULT_LAC = -1;
+    public static final long DEFAULT_CID = -1;
 }
